Tried to get lower absolute values for the e-powers, not working (yet).

+++ includes.hpp
@@ -2,6 +2,7 @@
 #define INCLUDES_H
 
 #include <algorithm>
+#include <cfenv>
 #include <chrono>
 #include <cmath>
 #include <csignal>
@@ -13,16 +14,18 @@
 #include <iostream>
 #include <sstream>
 #include <vector>
+ 
+//#pragma STDC FENV_ACCESS ON
 
 struct Node {
-   std::vector< double > weights;
-   double value;
-   double delta;
+   std::vector< long double > weights;
+   long double value;
+   long double delta;
 };
 
 struct OutputNode {
-   double value;
-   double delta;
+   long double value;
+   long double delta;
 };
 
 #endif
